Create Human Settlement Detection Challenge
Human settlements are rapidly expanding, creating an urgent need for efficient and accurate monitoring systems to track and manage this growth. Satellite imagery provides a comprehensive view of land usage, making it a valuable resource for detecting and analyzing human settlements. Currently, classification of human settlements from satellite imagery is often performed by experts, which is a demanding and time-consuming task. The result is often human-biased and may not scale effectively to handle large datasets. Thus, there is a pressing need for automated solutions that can accurately and efficiently identify and classify human settlements in satellite imagery.

The primary objective of this challenge is to develop a robust and accurate machine learning model that can detect human settlements in satellite imagery. The task is to build a binary classification model that can predict the presence of human settlements in the given test data. The evaluation of these predictions will determine the winner of the challenge. Participants are expected to submit their final model along with its predictions on the test dataset.

Developing an effective machine learning model for detecting human settlements in satellite imagery is crucial for several reasons. Firstly, it supports Sustainable Development Goal (SDG) 11, which aims to make cities and human settlements inclusive, safe, resilient, and sustainable. By accurately monitoring population growth and urban expansion, this technology can aid in urban planning, resource allocation, and disaster management. Additionally, it provides governments and organizations with the data necessary to address challenges related to infrastructure development, environmental conservation, and public health. Automated detection systems can handle large amounts of data, overcoming the limitations of manual classification which is often human-biased and labor-intensive. The solutions derived from this challenge will contribute to the broader efforts of managing human settlements sustainably in the face of rapid urbanization.

link : https://zindi.africa/competitions/inegi-gcim-human-settlement-detection-challenge